requirements.txt
APScheduler==3.6.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
pandas
Pillow==8.1.0
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.2
pytz==2021.1
requests==2.25.1
six==1.15.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
